+++ next.config.mjs
@@ -1,21 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {
-  images: {
-    remotePatterns: [
-      {
-        protocol: 'https',
-        hostname: 'lh3.googleusercontent.com',
-        port: '',
-        pathname: '/**',
-      },
-      {
-        protocol: 'https',
-        hostname: 's.gravatar.com',
-        port: '',
-        pathname: '/**',
-      },
-    ],
-  }
-};
+const nextConfig = {};
 
 export default nextConfig;