+++ src/app/api/auth/[auth0]/route.ts
@@ -1,6 +1,0 @@
-// Auth0 endpoint
-// source: https://auth0.com/docs/quickstart/webapp/nextjs/interactive
-import { handleAuth } from '@auth0/nextjs-auth0';
-
-export const GET = handleAuth();
-export const POST = handleAuth(); // work-around for AddToCart