+++ src/app/privacy/page.tsx
@@ -1,7 +1,0 @@
-export default async function PrivacyPage() {
-  return (
-    <div>
-      <h2>TODO</h2>
-    </div>
-  );
-}