+++ src/app/product/[productSlug]/page.tsx
@@ -1,7 +1,0 @@
-export default async function ProductPage() {
-  return (
-    <div>
-      <h2>TODO</h2>
-    </div>
-  );
-}