+++ src/components/Header.tsx
@@ -1,19 +1,34 @@
 'use client';
 
-import React from 'react';
-import { useUser } from '@auth0/nextjs-auth0/client';
+import SearchBar from './SearchBar';
+import { useShoppingCart } from 'use-shopping-cart';
 
 const Header: React.FC = () => {
-  const { user } = useUser();
+  const { cartCount } = useShoppingCart();
 
+  // TODO_AUTH_ACCOUNT
   return (
-    <header>
-      <a href="/">SwagSticker</a>
-      <div>
-        <p>Searchbar</p>
+    <header className="flex justify-between items-center px-4 py-2 bg-white shadow-md">
+      <div className="container mx-auto flex justify-between items-center">
+        <a href="/" className="text-xl font-bold">
+          SwagSticker
+        </a>
+        <SearchBar />
+        <div className="flex space-x-4">
+          {false ? (
+            <a href="/account" className="text-blue-500 hover:underline">
+              Account
+            </a>
+          ) : (
+            <a href="/api/auth/login" className="text-blue-500 hover:underline">
+              Sign in
+            </a>
+          )}
+          <a href="/cart" className="text-blue-500 hover:underline">
+            Cart ({cartCount})
+          </a>
+        </div>
       </div>
-      {user ? <a href="/account">Account</a> : <a href="/api/auth/login">Sign in</a>}
-      <a href="/cart">Cart</a>
     </header>
   );
 };