+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2024-10-14
+
+- Client-side cart item storage
+- Catalog filters/pagination
+- 47 real products with mockup images
+
+## [0.0.2] - 2024-10-14
+
+MVP landing page and product catalog
+
 ## [0.0.1] - 2024-09-18
 
 Boilerplate Next.js (App Router) code